refactor: move has_case_statuses() to case.py, eliminating import cycles
_helpers.py held two unrelated responsibilities: pure primitives (no
upward dependencies) and a case-aware predicate (has_case_statuses)
that needed VulnerabilityCase under a TYPE_CHECKING guard.  That guard
was the root of five static-analysis cycles reported against the
core/models/ package.

Move has_case_statuses() to case.py (where VulnerabilityCase lives),
remove the TYPE_CHECKING guard and from __future__ import annotations
from _helpers.py, and update both production callers and the three
test inline imports.  pyright reports 0 errors / 0 warnings on
vultron/core/models/; both unit and integration suites pass.

Closes #1933
